% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_sqlite.R
\name{append_to_sqlite}
\alias{append_to_sqlite}
\title{Callback function that appends new sections to the SQLite table}
\usage{
append_to_sqlite(con, table_name, date_cols, datetime_cols)
}
\arguments{
\item{datetime_cols}{name of columns containint POSIXt objects}

\item{x}{data frame}

\item{data_cols}{name of columns containing Date objects}
}
\description{
Callback function that appends new sections to the SQLite table
}
