% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquachem_measurement.R
\name{aquachem_measurement}
\alias{aquachem_measurement}
\title{Get data from Aquachem measurement table}
\usage{
aquachem_measurement(dbase_connection, code_name, sampling_date,
  reportable)
}
\arguments{
\item{code_name}{(string) code name of the sampling}

\item{sampling_date}{(string) date of the sampling}

\item{reportable}{(string) 0 or 1, 1 if the output is part of the report}

\item{my_connection}{A valid connection to Aquachem database.}
}
\value{
A tibble (tidyverse data.frame).
}
\description{
Get all LIMS data linked to a single sampling. Filter options are the location of 
the fieldsampling and the date of the sampling
}
