% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_inbo_dbase.R
\name{connect_inbo_dbase-defunct}
\alias{connect_inbo_dbase-defunct}
\title{Connect to an INBO database}
\usage{
connect_inbo_dbase(database_name)
}
\arguments{
\item{database_name}{char Name of the INBO database you want to connect}
}
\value{
odbc connection
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
Connects to an INBO database by simply providing the database's name as an
argument.
The function can only be used from within the INBO network.

For more information, refer to
\href{https://inbo.github.io/tutorials/tutorials/r_database_access/}{this tutorial}.
}
\examples{
\dontrun{
connection <- connect_inbo_dbase("D0021_00_userFlora")
connection <- connect_inbo_dbase("W0003_00_Lims")
}

}
\seealso{
\code{\link{inborutils-defunct}}
}
\keyword{internal}
