% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_inbo_dbase.R
\name{connect_inbo_dbase-deprecated}
\alias{connect_inbo_dbase-deprecated}
\title{Connect to an INBO database}
\usage{
connect_inbo_dbase(database_name)
}
\arguments{
\item{database_name}{char Name of the INBO database you want to connect}
}
\value{
odbc connection
}
\description{
Connects to an INBO database by simply providing the database's name as an
argument.
The function can only be used from within the INBO network.
}
\details{
For more information, refer to
\href{https://inbo.github.io/tutorials/tutorials/r_database_access/}{this tutorial}.
}
\examples{
\dontrun{
connection <- connect_inbo_dbase("D0021_00_userFlora")
connection <- connect_inbo_dbase("W0003_00_Lims")
}

}
\seealso{
\code{\link{inborutils-deprecated}}
}
\keyword{internal}
