% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_inbo_dbase.R
\name{connect_inbo_dbase}
\alias{connect_inbo_dbase}
\title{Connect to an INBO database}
\usage{
connect_inbo_dbase(database_name, sql_driver = "SQL Server")
}
\arguments{
\item{database_name}{char Name of the INBO database you want to connect}

\item{sql_driver}{char Driver (\code{SQL Server} by default) to use for connection}
}
\value{
odbc connection
}
\description{
Connect to an INBO database
}
\examples{
connection <- connect_inbo_dbase("D0021_00_userFlora")
}
