% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_df_sf.R
\name{convertdf_enc}
\alias{convertdf_enc}
\title{Convert encoding of character and factor variables in a dataframe}
\usage{
convertdf_enc(x, to = "UTF-8")
}
\arguments{
\item{x}{A dataframe.}

\item{to}{Encoding string to convert to.
All \code{R} platforms support \code{""} (for the encoding of the current
locale), \code{"latin1"} and \code{"UTF-8"}.
See \code{\link[base]{iconv}} for more information.}
}
\value{
The original dataframe, with character variables (and levels of
(character) factor variables) converted to the specified encoding.
}
\description{
Convert encoding of character and factor variables in a dataframe
}
