% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coordinate_example}
\alias{coordinate_example}
\title{Example data.frame with coorindates}
\format{
A data frame with 52 rows and 3 variables:
\itemize{
\item{id: resource identifier}
\item{latitude: Latitude of the coordinates}
\item{longitude: Longitude of the coordinates}
}
}
\usage{
coordinate_example
}
\description{
A dataset containing 52 coordinates as latitude and longitude
}
\keyword{datasets}
