% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_pts_in_polygons.R
\name{counts_pts_in_polygons}
\alias{counts_pts_in_polygons}
\title{Count points contained in polygons}
\usage{
counts_pts_in_polygons(df, polygons_df, lat_col = "decimalLatitude",
  lon_col = "decimalLongitude", crs = "+init=epsg:4326", verbose = FALSE)
}
\arguments{
\item{df}{df containing geographical information (lat-log cols) about points}

\item{lat_col}{name of the column containing latitude}

\item{lon_col}{name of the column containing longitude}
}
\value{
a SpatialPolygonDataframe with @data df containing point counts
}
\description{
The function takes longitude-latitude columns from df to create
a SpatialPointsDataframe ane make a spatial join over the
SpatialPolygonsDataFrame in order to count how many points are in each
polygon. A previous CRS transformation of the polygon's CRS is performed
in order to match the CRS given as input.
}
