% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_inbo_dbase.R
\docType{methods}
\name{dbDisconnect,OdbcConnection-method}
\alias{dbDisconnect,OdbcConnection-method}
\title{Overwrite the odbc function}
\usage{
\S4method{dbDisconnect}{OdbcConnection}(conn, ...)
}
\arguments{
\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[=dbConnect]{dbConnect()}}.}

\item{...}{Other parameters passed on to methods.}
}
\description{
Overwrite the odbc function
}
