% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_factors_to_char.R
\name{df_factors_to_char}
\alias{df_factors_to_char}
\title{Convert all factors of a dataframe to characters}
\usage{
df_factors_to_char(rp)
}
\arguments{
\item{rp}{A dataframe}
}
\description{
The function checks the which columns are factors and converts these to
character vectors
}
\details{
all credits to Thierry Onkelinx
}
\examples{
df_factors_to_char(PlantGrowth) # column group will be chars as well

}
\seealso{
Other Data_handling_utilities: 
\code{\link{convertdf_enc}()},
\code{\link{csv_to_sqlite}()}
}
\concept{Data_handling_utilities}
