% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/florabank_queries.R
\name{florabank_observations}
\alias{florabank_observations}
\title{Get all validated observations for a taxon from the florabank database}
\usage{
florabank_observations(scient_name, dutch_name)
}
\arguments{
\item{scient_name}{(Part of) a scientific taxon name}

\item{dutch_name}{(Part of) a Dutch taxon name, may contain wildcards}
}
\value{
A dataframe with the following variables: "NaamNederlands",
"NaamWetenschappelijk", "Bron", "BeginDatum", "EindDatum", "hok",
"Toponiem", "CommentaarTaxon", "CommentaarHabitat", "Voornaam", "Achternaam",
"ID", "X_waarneming", "Y_waarneming", "X_meting", "Y_meting"
}
\description{
This function takes as input either (part of) a scientific name or (part of)
the Dutch name for a taxon, queries the florabank, and returns a dataframe
observation level information about the matching taxa. The following
wildcards may be used within the input parameters:
\code{\%}: Represents zero or more characters;
\code{_}: Represents a single character;
\code{[]}: Represents any single character within the brackets, e.g. \code{[sp]}
\code{^}: Represents any character not in the brackets, e.g. \code{[^sp]}
\code{-}: Represents a range of characters, e.g. \code{[a-z]}
}
\examples{
\dontrun{
Dactmacu1 <-	florabank_observations(scient_name = 'Dactylorhiza maculata (L.) Soó')
Dactmacu2 <-	florabank_observations(dutch_name = 'Gevlekte orchis')
all.equal(Dactmacu1, Dactmacu2)
# use wildcards to retrieve all partial matches to a name
Dactmacu3 <-	florabank_observations(scient_name = 'Dactylorhiza maculata\%')
}
}
