% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coverage_wcs.R
\name{get_coverage_wcs}
\alias{get_coverage_wcs}
\title{Get a layer from a web coverage service within a bounding box}
\usage{
get_coverage_wcs(
  wcs = c("dtm", "dsm", "omz", "omw"),
  bbox,
  layername,
  resolution,
  wcs_crs = "EPSG:4258",
  output_crs = "EPSG:31370",
  bbox_crs = "EPSG:31370",
  version = c("1.0.0", "2.0.1"),
  ...
)
}
\arguments{
\item{wcs}{One of \code{"dtm"}, \code{"dsm"}, \code{"omz"}, \code{"omw"}}

\item{bbox}{An object of class bbox of length 4.}

\item{layername}{Character string; name of the layer}

\item{resolution}{Output resolution in meters}

\item{wcs_crs}{Native CRS in which the raster layers are stored on the \code{WCS}}

\item{bbox_crs}{CRS in which bbox coordinates are passed}

\item{version}{\code{WCS} version to be used.}

\item{...}{Additional key-value pairs passed on to the \code{WCS} query}

\item{output_wcs}{Output CRS.
May involve reprojection.}
}
\value{
A \code{SpatRaster} object
}
\description{
The function sends a query to a WCS service, downloads it to a temporary file
from which it is read with \code{terra::rast()} - if needed reprojected -
and returned as a \code{SpatRaster}object
}
\details{
The following WCS services can currently be used:
\itemize{
\item \code{"omz"}: orthophotomosaic summer images Flanders
\item \code{"omw"}: orthophotomosaic winter images Flanders
\item \code{"dtm"}: digital terrain model Flanders
\item \code{"dsm"}: digital surface model Flanders
See https://metadata.vlaanderen.be/metadatacenter/srv/dut/catalog.search#/search?keyword=OGC:WCS for more information # nolint
}
}
\examples{
\dontrun{
 bbox <- sf::st_bbox(
   c(xmin = 155800 , xmax = 155850, ymin = 132700, ymax = 132750),
     crs = sf::st_crs(31370))
 get_coverage_wcs(wcs = "dsm",
   bbox = bbox,
   layername = "EL.GridCoverage.DSM",
   resolution = 1)
}

}
\seealso{
Other GIS_utilities: 
\code{\link{get_feature_wfs}()},
\code{\link{guess_crs}()},
\code{\link{plot_coordinates_on_map}()},
\code{\link{transform_coordinates}()}
}
\concept{GIS_utilities}
