% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature_wfs.R
\name{get_feature_wfs}
\alias{get_feature_wfs}
\title{Get a layer from a web feature service}
\usage{
get_feature_wfs(
  wfs,
  version = "2.0.0",
  layername,
  crs,
  bbox = NULL,
  filter = NULL,
  cql_filter = NULL,
  output_format = NULL,
  property_name = NULL,
  result_type = c("results", "hits"),
  ...
)
}
\arguments{
\item{wfs}{Web address for the service which you want to query features from}

\item{version}{Version number for the service.
For instance \code{"2.0.0"}.}

\item{layername}{the name of a layer hosted by the web feature service}

\item{crs}{coordinate reference system to represent the features.
For instance \code{"EPSG:31370"}.}

\item{bbox}{Optional bounding box.
Pass this as a named vector with names "xmin", "xmax", "ymin", "ymax".}

\item{filter}{Optional \link{standard OGC filter}
(https://www.ogc.org/standards/filter) specification}

\item{cql_filter}{Optional \link{Contextual Query Language}
(https://portal.ogc.org/files/96288) filter.
This currently only works if the \code{WFS} is hosted on a \code{GeoServer}.}

\item{output_format}{Optional output format supported by the \code{WFS}.}

\item{property_name}{Optional character string.
Which fields or columns to return?
If you want to specify multiple columns, separate them by a comma.
The column containing the feature geometry is usually called \code{geom},
\code{geometry} or \code{SHAPE}.}

\item{result_type}{For version \code{"2.x.x"}, this can be either \code{"results"}
(default) or \code{"hits"}.
The former returns the requested features, the latter returns the number of
requested features.}

\item{...}{Additional key-value pairs passed on to the WFS query.}
}
\description{
This function constructs a \code{URL} request from its arguments and either reads
in the resulting vector layer as a \code{sf} object or returns the number of
features that are requested.
The request is made up of key-value pairs and additional key-value pairs can
be passed to the function.
The full documentation for the \code{WFS} standard can be consulted from
https://www.ogc.org/standards/wfs
}
\details{
See https://inbo.github.io/tutorials/tutorials/spatial_wfs_services/
for more information.
}
\examples{
\dontrun{
vlaanderen <- get_feature_wfs(
wfs = paste0("https://eservices.minfin.fgov.be/",
             "arcgis/services/R2C/Regions/MapServer/WFSServer"),
layername = "regions",
crs = "EPSG:31370",
filter = paste0("<Filter><PropertyIsEqualTo><PropertyName>",
                "regions:NameDUT</PropertyName><Literal>'Vlaams Gewest'",
                "</Literal></PropertyIsEqualTo></Filter>"))
}
}
\seealso{
Other GIS_utilities: 
\code{\link{get_coverage_wcs}()},
\code{\link{guess_crs}()},
\code{\link{plot_coordinates_on_map}()},
\code{\link{transform_coordinates}()}
}
\concept{GIS_utilities}
