% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_species_name_match.R
\name{get_name_gbif}
\alias{get_name_gbif}
\title{Helper function for retrieving informations from GBIF Taxonomy Backbone}
\usage{
get_name_gbif(...)
}
\description{
Helper function for retrieving informations from GBIF Taxonomy Backbone
}
\keyword{internal}
