% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_crs.R
\name{guess_crs}
\alias{guess_crs}
\title{Take a dataset and plot in different coordinate reference systems (CRS) on a
leaflet map to check the expected CRS}
\usage{
guess_crs(
  df,
  col_x,
  col_y,
  belgium = TRUE,
  crs_try = c("EPSG:4326", "EPSG:31370", "EPSG:28992", "EPSG:32631", "EPSG:3812",
    "EPSG:3035")
)
}
\arguments{
\item{df}{data.frame with a x and y coordinate column}

\item{col_x}{(character) name of the x (longitude) column}

\item{col_y}{(character) name of the y (latitude) column}

\item{belgium}{(logical) If \code{TRUE}, coordinates are expected to be in Belgium}

\item{crs_try}{(list) \code{EPSG} codes of the different CRS to evaluate
By default, the following six CRS are added to the map:
\itemize{
\item{"EPSG:4326"}{WGS 84,
https://spatialreference.org/ref/epsg/wgs-84/}
\item{"EPSG:31370"}{Belge 1972/Belgian Lambert 72,
https://spatialreference.org/ref/epsg/31370/}
\item{"EPSG:28992"}{Amersfoort/Rijksdriehoek nieuw,
https://spatialreference.org/ref/epsg/28992/}
\item{"EPSG:32631"}{WGS 84 / UTM zone 31N,
https://spatialreference.org/ref/epsg/32631/}
\item{"EPSG:3812"}{ETRS89/Belgian Lambert 2008,
https://spatialreference.org/ref/epsg/3812/}
\item{"EPSG:3035"}{ETRS89 / ETRS-LAEA,
https://spatialreference.org/ref/epsg/3035/}
}}
}
\description{
When retrieving a number of data-points, but the kind of CRS is not
provided/known (for any reason), this utility plots the data for different
CRS on a map to make comparison possible. The main function is
\code{guess_crs} which creates the map with different options.
Custom projections can be provided by the user as well.
}
\details{
For each of the given CRS, the coordinates columns are given the
the specified CRS In a next step, these CRS are converted to
\code{wgs84} and plotted on a \code{openstreetmap} background with leaflet.
The interactive map provides the possibility to select/unselect specific
layers.
}
\examples{
\dontrun{
guess_crs(data, "x", "y")
}

}
\seealso{
crsuggest::guess_crs() for a (better) alternative

Other GIS_utilities: 
\code{\link{plot_coordinates_on_map}()},
\code{\link{transform_coordinates}()}
}
\concept{GIS_utilities}
