% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_projection.R
\name{guess_projection}
\alias{guess_projection}
\title{Take a dataset and plot in different projections on a leaflet map to
check the expected projection}
\usage{
guess_projection(
  df,
  col_x,
  col_y,
  belgium = TRUE,
  projections = c("EPSG:4326", "EPSG:31370", "EPSG:28992", "EPSG:32631", "EPSG:3812",
    "EPSG:3035")
)
}
\arguments{
\item{df}{data.frame with a x and y coordinate column}

\item{col_x}{(char) name of the x (longitude) column}

\item{col_y}{(char) name of the y (latitude) column}

\item{belgium}{(bool) If TRUE, coordinates are expected to be in Belgium}

\item{projections}{(list) EPSG codes of the different projections to evaluate
By default, the following six projections are added to the map:
\itemize{
\item{"EPSG:4326"}{WGS 84,
http://spatialreference.org/ref/epsg/wgs-84/}
\item{"EPSG:31370"}{Belge 1972/Belgian Lambert 72,
http://spatialreference.org/ref/epsg/31370/}
\item{"EPSG:28992"}{Amersfoort/Rijksdriehoek nieuw,
http://spatialreference.org/ref/epsg/28992/}
\item{"EPSG:32631"}{WGS 84 / UTM zone 31N,
http://spatialreference.org/ref/epsg/32631/}
\item{"EPSG:3812"}{ETRS89/Belgian Lambert 2008,
http://spatialreference.org/ref/epsg/3812/}
\item{"EPSG:3035"}{ETRS89 / ETRS-LAEA,
http://spatialreference.org/ref/epsg/3035/}
}}
}
\description{
When retrieving a number of data-points, but the kind of projection is not
provided/known (for any reason), this utility plots the data for different
projections on a map to make comparison possible. The main function is
\code{guess_projection } which creates the map with different options.
Custom projections can be provided by the user as well.
}
\details{
For each of the given projections, the coordinates columns are given the
the specified projection. In a next step, these projections are converted to
wgs84 and plotted on a openstreetmap background with leaflet. The interactive
map provides the possibility to select/unselect specific layers.
}
\examples{
\dontrun{
guess_projection(data, "x", "y")
}

}
\seealso{
Other GIS_utilities: 
\code{\link{plot_coordinates_on_map}()},
\code{\link{reproject_coordinates}()},
\code{\link{transform_coordinates}()}
}
\concept{GIS_utilities}
