% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_zenodo.R
\name{human_filesize}
\alias{human_filesize}
\title{Human-readable binary file size}
\usage{
human_filesize(x)
}
\arguments{
\item{x}{An integer, i.e. the number of bytes (B).}
}
\value{
A string.
}
\description{
Takes an integer (referring to number of bytes) and returns an optimally
human-readable
\href{https://en.wikipedia.org/wiki/Binary_prefix}{binary-prefixed}
byte size (KiB, MiB, GiB, TiB, PiB, EiB)
}
\keyword{internal}
