% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_inbo_dbase.R, R/inborutils_defunct.R
\name{connect_inbo_dbase}
\alias{connect_inbo_dbase}
\alias{inborutils-defunct}
\title{Defunct functions in package \pkg{inborutils}}
\usage{
connect_inbo_dbase(database_name)
}
\description{
The functions listed below are moved to package \pkg{inbodb}
and removed from \pkg{inborutils}. Please install package
\pkg{inbodb} to continue using these functions (instructions on
https://inbo.github.io/inbodb/).
}
\section{connect_inbo_dbase}{

For \code{connect_inbo_dbase}, use
\href{https://inbo.github.io/inbodb/reference/connect_inbo_dbase.html}{inbodb::connect_inbo_dbase()}
}

\keyword{internal}
