% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inborutils-deprecated.R
\name{inborutils-deprecated}
\alias{inborutils-deprecated}
\title{Deprecated functions in inborutils}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
\item \code{\link{reproject_coordinates}}: This function is deprecated,
and will be removed in the next version of this package.
}
}
