% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inboveg_layer_coverage.R
\name{inboveg_coverage}
\alias{inboveg_coverage}
\title{Taxa and coverage per vegetation layer}
\usage{
inboveg_coverage(connection, recording_type = c("Classic",
  "Classic-emmer", "Classic-ketting"), survey_name)
}
\arguments{
\item{connection}{odb DBIConnection-class, using DBI-ODBC like connectors}

\item{recording_type}{vector of char, default c('Classic', 'Classic-emmer',
'Classic-ketting')}

\item{survey_name}{vector of char or NULL (default NULL),
i.e. all surveys are taken into account}
}
\value{
data.frame
}
\description{
For the requested surveys, the coverage for each taxon and each vegetation
layer is provided.
}
\examples{
\dontrun{
con <- connect_inbo_dbase("D0010_00_Cydonia")
coverage <- inboveg_coverage(con, survey_name = c('Sigma_LSVI_2012'))
coverage <- inboveg_coverage(con, recording_type = c('Classic',
    'Classic-emmer', 'Classic-ketting'), survey_name = c('Sigma_LSVI_2012'))
 dbDisconnect(con)
}
}
\seealso{
Other inboveg: \code{\link{inboveg_classification}},
  \code{\link{inboveg_header}},
  \code{\link{inboveg_metadata_kop}},
  \code{\link{inboveg_recordings}},
  \code{\link{inboveg_survey}},
  \code{\link{inboveg_synonym_list}}
}
\concept{inboveg}
