% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inboveg_header.R
\name{inboveg_header-defunct}
\alias{inboveg_header-defunct}
\title{Query header information from INBOVEG}
\usage{
inboveg_header(connection, survey_name, rec_type, multiple = FALSE,
  collect = FALSE)
}
\arguments{
\item{survey_name}{A character string or a character vector
giving the name or names of the survey(s) for which you want to extract
header information.
If missing, all surveys are returned.}

\item{rec_type}{A character vector giving the name of record type for which
you want to extract header information e.g. 'Classic', 'Classic-emmer',
'Classic-ketting', 'BioHab', 'ABS'. If missing, all recording types are
returned.}

\item{connection}{dbconnection with the database 'Cydonia'
on the inbo-sql07-prd server}

\item{multiple}{If TRUE, survey_name can take a character vector with
multiple survey names that must match exactly. If FALSE (the default),
survey_name must be a single character string (one survey name) that can
include wildcards to allow partial matches}

\item{collect}{If FALSE (the default), a remote tbl object is returned.
This is like a reference to the result of the query but the full result of
the query is not brought into memory. If TRUE the full result of the query is
collected (fetched) from the database and brought into memory of the working
environment.}
}
\value{
A remote tbl object (collect = FALSE) or a tibble dataframe (collect
= TRUE) with variables RecordingGivid, Name, UserReference, Observer,
LocationCode, Latitude, Longitude, Area (in m2), Length (in cm), Width
(in cm), VagueDateType, VagueDateBegin, VagueDateEnd, SurveyId, RecTypeID.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
This function queries the INBOVEG database for header
information (metadata for a vegetation-recording or releve) for one or more
surveys and the recorder type. See the examples for how to get information
for all surveys.
}
\examples{
\dontrun{
con <- connect_inbo_dbase("D0010_00_Cydonia")

# get header information from a specific survey and a specific recording type
# and collect the data
header_info <- inboveg_header(con,
  survey_name = "OudeLanden_1979",
  rec_type = "Classic", collect = TRUE
)

# get header information from several specific surveys
header_severalsurveys <- inboveg_header(con,
  survey_name =
    c("MILKLIM_Heischraal2012", "NICHE Vlaanderen"), multiple = TRUE
)

# get header information of all surveys,  don't collect the data
all_header_info <- inboveg_header(con)

# close the connection when done
dbDisconnect(con)
rm(con)
}

}
\seealso{
\code{\link{inborutils-defunct}}

Other inboveg: 
\code{\link{inboveg_classification-defunct}},
\code{\link{inboveg_qualifiers-defunct}},
\code{\link{inboveg_recordings-defunct}},
\code{\link{inboveg_survey-defunct}}
}
\concept{inboveg}
\keyword{internal}
