% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inboveg_header.R
\name{inboveg_header}
\alias{inboveg_header}
\title{Query header information from INBOVEG}
\usage{
inboveg_header(connection, survey_name, rec_type, collect = FALSE)
}
\arguments{
\item{connection}{dbconnection with the database 'Cydonia'
on the inbo-sql07-prd server}

\item{survey_name}{A character vector giving the name of the survey for which
you want to extract header information. If missing, all surveys are returned.}

\item{rec_type}{A character vector giving the name of record type for which
you want to extract header information e.g. 'Classic', 'Classic-emmer',
'Classic-ketting', 'BioHab', 'ABS'. If missing, all recording types are returned.}

\item{collect}{If FALSE (the default), a remote tbl object is returned.
This is like a reference to the result of the query but the full result of
the query is not brought into memory. If TRUE the full result of the query is
collected (fetched) from the database and brought into memory of the working
environment.}
}
\value{
A remote tbl object (collect = FALSE) or a tibble dataframe (collect
= TRUE) with variables RecordingGivid, Name, UserReference, LocationCode,
Latitude, Longitude, Area (in m2), Length (in cm), Width (in cm), SurveyId, RecTypeID.
}
\description{
This function queries the INBOVEG database for header
information (metadata for a vegetation-relevé) for one survey by the name
of the survey and the recorder type. See the examples for how to get
information for all surveys.
}
\examples{
\dontrun{
con <- connect_inbo_dbase("D0010_00_Cydonia")

# get header information from a specific survey and a specific recording type
and collect the data
header_info <- inboveg_header(con, survey_name = "OudeLanden_1979",
rec_type = "Classic", collect = TRUE)

# get header information of all surveys,  don't collect the data
all_header_info <- inboveg_header(con)

# Close the connection when done
dbDisconnect(con)
rm(con)
}
}
\seealso{
Other inboveg: 
\code{\link{inboveg_classification}()},
\code{\link{inboveg_qualifiers}()},
\code{\link{inboveg_recordings}()},
\code{\link{inboveg_survey}()}
}
\concept{inboveg}
