% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inboveg_recording_metadata.R
\name{inboveg_metadata_kop}
\alias{inboveg_metadata_kop}
\title{Provide metadata about the vegetation recordings of the INBOVEG database}
\usage{
inboveg_metadata_kop(connection, recording_type = c("Classic",
  "Classic-emmer", "Classic-ketting"), survey_name)
}
\arguments{
\item{connection}{odb DBIConnection-class, using DBI-ODBC like connectors}

\item{recording_type}{vector of char, default c('Classic', 'Classic-emmer',
'Classic-ketting')}

\item{survey_name}{vector of char or NULL (default NULL),
i.e. all surveys are taken into account}
}
\value{
data.frame
}
\description{
For each vegetation recording identified by RecordingGivid, the metadata is
provided in the resulting output table
}
\examples{
\dontrun{
library(DBI)
con <- connect_inbo_dbase("D0010_00_Cydonia")
metadata <- inboveg_metadata_kop(con)
metadata <- inboveg_metadata_kop(con, survey_name = c("Sigma_LSVI_2012"))
dbDisconnect(con)
}
}
