% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inboveg_survey.R
\name{inboveg_survey-deprecated}
\alias{inboveg_survey-deprecated}
\title{Query survey information from INBOVEG}
\usage{
inboveg_survey(connection, survey_name, collect = FALSE)
}
\arguments{
\item{survey_name}{A character vector giving the names of the surveys for
which you want to extract survey information.}

\item{connection}{dbconnection with the database 'Cydonia'
on the inbo-sql07-prd server}

\item{collect}{If FALSE (the default), a remote tbl object is returned. This
is like a reference to the result of the query but the full result of the
query is not brought into memory. If TRUE the full result of the query is
collected (fetched) from the database and brought into memory of the working
environment.}
}
\value{
A remote tbl object (collect = FALSE) or a tibble dataframe (collect
= TRUE) with variables Id, Name, Description, Owner and Creator.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function queries the INBOVEG database for survey information
(metadata about surveys) for one or more survey(s) by the name of the survey.
See the examples for how to get information for all surveys.
}
\examples{
\dontrun{
con <- connect_inbo_dbase("D0010_00_Cydonia")

# get information of a specific survey and collect data
survey_info <- inboveg_survey(con, survey_name = "OudeLanden_1979",
collect = TRUE)

# get information of all surveys and collect data
allsurveys <- inboveg_survey(con)

# Close the connection when done
dbDisconnect(con)
rm(con)
}

}
\seealso{
\code{\link{inborutils-deprecated}}

Other inboveg: 
\code{\link{inboveg_classification-deprecated}},
\code{\link{inboveg_header-deprecated}},
\code{\link{inboveg_qualifiers-deprecated}},
\code{\link{inboveg_recordings-deprecated}}
}
\concept{inboveg}
\keyword{internal}
