% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inboveg_synonym.R
\name{inboveg_synonym_list}
\alias{inboveg_synonym_list}
\title{Provide accepted names list from INBOVEG database}
\usage{
inboveg_synonym_list(connection, TaxonListGIVID = "TL2011092815101010",
  ListName = "INBO-2011 Sci")
}
\arguments{
\item{connection}{odb DBIConnection-class, using DBI-ODBC like connectors}

\item{TaxonListGIVID}{char taxon list}

\item{ListName}{char preferred name list}
}
\value{
data.frame
}
\description{
Get all taxon names available from an INBOVEG database taxon list
(identified with the \code{TaxonListGIVID}) and give for each name the corresponding
preferred taxon name as available in the specified reference list
(identified by \code{ListName})
}
\examples{
\dontrun{
con <- connect_inbo_dbase("D0013_00_Futon")
plant_overview <- inboveg_synonym_list(con,
                           TaxonListGIVID = 'TL2011092815101010',
                           ListName = 'INBO-2011 Sci')
dbDisconnect(con)
}

}
