% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mow_data.R
\name{mow_header_split}
\alias{mow_header_split}
\title{Get data entry from MOW header data file}
\usage{
mow_header_split(inputline)
}
\arguments{
\item{inputline}{character string coming from MOW data file}
}
\value{
char string with the entry
}
\description{
Get data entry from MOW header data file
}
