% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_inbo_dbase.R
\name{on_connection_opened}
\alias{on_connection_opened}
\title{Rstudio Viewer integration}
\usage{
on_connection_opened(connection, code)
}
\arguments{
\item{connection}{odbc connection}

\item{code}{dbase connection code}
}
\description{
Rstudio Viewer integration
}
