% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coordinates_on_map.R
\name{plot_coordinates_on_map}
\alias{plot_coordinates_on_map}
\title{plot x/y coordinates on a map, from given projection}
\usage{
plot_coordinates_on_map(df, col_long, col_lat, projection)
}
\arguments{
\item{df}{data.frame with a x and y coordinate column}

\item{col_long}{(char) name of the x (longitude) column}

\item{col_lat}{(char) name of the y (latitude) column}

\item{projection}{projection string of class CRS-class defining the
current projection}
}
\value{
leaflet map with coordinates added as dots
}
\description{
This function first converts the x/y coordinates of the given the data.frame
to GWS84 in order to map the data with leaflet
}
