% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_label_splitter.R
\name{plot_label_splitter}
\alias{plot_label_splitter}
\title{Split a long (plot) label into a two-liner}
\usage{
plot_label_splitter(label, maxlength)
}
\arguments{
\item{label}{A character array (the label)}

\item{maxlength}{When the label exceeds this length, it will be splitted into
a two-line label}
}
\description{
When a label exceeds the maximum length as defined by the user, the function
splits the text/label on the first space after the center of the string. This
is very useful for long labels in a plot
}
\details{
The function can be easily applied to a data.frame by using e.g. sapply
}
\examples{
plot_label_splitter("Exotic long label", 10)

}
