% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rain_knmi_2012}
\alias{rain_knmi_2012}
\title{Example \code{data.frame} with \code{KNMI} downloaded data}
\format{
A \code{data.frame} with 1536 rows and 9 variables:
\itemize{
\item{\code{value}: measured value}
\item{\code{datetime}: \code{datetime} of the measurement}
\item{\code{unit}: unit (mm)}
\item{\code{variable_name}: precipitation}
\item{\code{longitude}: coordinate}
\item{\code{latitude}: coordinate}
\item{\code{location_name}: station name}
\item{\code{source_filename}: filename from which the data was read}
\item{\code{quality_code}: empty string as \code{KNMI} does not provide this}
}
}
\usage{
rain_knmi_2012
}
\description{
A dataset containing the rainfall from January 1st till February 1st for
\code{Vlissingen} and \code{Westdorpe}, as downloaded from \code{KNMI}
}
\seealso{
Other datasets: 
\code{\link{coordinate_example}},
\code{\link{species_example}}
}
\concept{datasets}
\keyword{datasets}
