% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_knmi_data.R
\name{read_knmi_data}
\alias{read_knmi_data}
\title{Read \code{KNMI} datafile (precipitation) as \code{data.frame}}
\usage{
read_knmi_data(filename, n_max = Inf)
}
\arguments{
\item{filename}{path/filename of the KMI datafile}

\item{n_max}{integer; shortcut for testing, loading just a section of the
data}
}
\value{
\code{data.frame} with the headers \code{location_name}, \code{datetime} (UTC),
\code{value}, \code{unit}, \code{variable_name}, \code{latitude}, \code{longitude} and
\code{source_filename}
}
\description{
Remark that this function is specifically to extract the precipitation (RH)
data, coordinates are extracted from the header of the \code{KNMI} data file
}
\examples{
\dontrun{
# see vignettes for more examples
file_path <- "knmi_file.txt"
knmi_data <- read_knmi_data(file_path)
}
}
\seealso{
Other download_functions: 
\code{\link{download_knmi_data_hour}()},
\code{\link{download_zenodo}()},
\code{\link{extract_soil_map_data}()},
\code{\link{read_kmi_data}()},
\code{\link{read_kml_file}()},
\code{\link{read_mow_data}()}
}
\concept{download_functions}
