% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_knmi_rainfall.R
\name{read_knmi_rainfall}
\alias{read_knmi_rainfall}
\title{Read KNMI rainfall datafile as data.frame}
\usage{
read_knmi_rainfall(filename, n_max = Inf)
}
\arguments{
\item{filename}{path/filename of the KMI datafile}

\item{n_max}{int shortcut for testing, loading just a section of the data}
}
\value{
data.frame with the headers location_name, datetime (UTC), value,
unit,variable_name, latitude, longitude and source_filename
}
\description{
Remark that this function is specifically to extract the rainfall (RH) data,
coordinates are extracted from the header of the KNMI data file
}
