% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mow_data.R
\name{read_mow_data}
\alias{read_mow_data}
\title{Read MOW data}
\usage{
read_mow_data(filename, n_max = Inf)
}
\arguments{
\item{filename}{path/filename of the RWS datafile}

\item{n_max}{int shortcut for testing, loading just a section of the data}
}
\value{
data.frame with the headers location_name, datetime (UTC), value,
unit,variable_name, latitude, longitude and source_filename
}
\description{
Coordinate data is contained in the header of the file
}
\examples{
\dontrun{
# see vignettes for more examples
file_path <- "mow_file.txt"
mow_data <- read_mow_data(file_path)
}
}
\seealso{
Other download_functions: 
\code{\link{download_knmi_data_hour}()},
\code{\link{download_zenodo}()},
\code{\link{extract_soil_map_data}()},
\code{\link{read_kmi_data}()},
\code{\link{read_kml_file}()},
\code{\link{read_knmi_data}()}
}
\concept{download_functions}
