% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rws_data.R
\name{read_rws_data}
\alias{read_rws_data}
\title{Load RWS data}
\usage{
read_rws_data(filename, n_max = Inf)
}
\arguments{
\item{filename}{path/filename of the RWS datafile}

\item{n_max}{int shortcut for testing, loading just a section of the data}
}
\value{
data.frame with the headers location_name, datetime (UTC), value,
unit,variable_name, latitude, longitude and source_filename
}
\description{
Currently, the function supports conductivity and water temperature data
files
}
