% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproject_coordinates.R
\name{reproject_coordinates}
\alias{reproject_coordinates}
\title{reproject XY coordinates from dframe columns}
\usage{
reproject_coordinates(df, col_long, col_lat, crs_input, crs_output)
}
\arguments{
\item{df}{data.frame with a x and y coordinate column}

\item{col_long}{(char) name of the x (longitude) column}

\item{col_lat}{(char) name of the y (latitude) column}

\item{crs_input}{projection string of class CRS-class defining the
current projection}

\item{crs_output}{projection string of class CRS-class defining the
projection to convert to}
}
\value{
data.frame with the same columns, but adapted coordinates for the
x and y column values
}
\description{
reproject XY coordinates from dframe columns
}
