% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_codingclub_session.R
\name{setup_codingclub_session}
\alias{setup_codingclub_session}
\title{Download files (code and data) for a specific coding club session}
\usage{
setup_codingclub_session(session_date)
}
\arguments{
\item{session_date}{The date of the coding-club session, in the "YYYYMMDD" format.}
}
\description{
This function will populate the content of /src and /data directories
for a specific coding club session whose date is passed as a parameter.
}
