% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coverage_wcs.R
\name{unpack_mht}
\alias{unpack_mht}
\title{Unpack or extract the \code{tif} file part from an \code{mht} file}
\usage{
unpack_mht(path)
}
\arguments{
\item{path}{A path to the \code{mht} file}
}
\description{
This helper function is needed on some \code{WCS} services from which an \code{mht}
file is downloaded rather than a \code{tif} file
}
\details{
Need three ways to read in the \code{mht} file to get the \code{tif} file out.
\code{read_lines()} cannot read all lines due to embedded \code{nulls}.
Therefore, also \code{read_lines_raw()} needed for positioning of \code{tif} part in
file.
\code{write_lines()} does not work correctly on \code{lines_raw[start:end]}
possibly a bug or edge case in \code{write_lines()}
Therefore, also \code{read_file_raw()} needed to extract from the raw vector
}
\keyword{internal}
