% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_inflation.R
\name{vif}
\alias{vif}
\alias{vif.lm}
\alias{vif.data.frame}
\alias{vif.default}
\title{Variance inflation factor}
\usage{
vif(mod, ...)

\method{vif}{lm}(mod, ...)

\method{vif}{data.frame}(mod, ...)

\method{vif}{default}(mod, ...)
}
\arguments{
\item{mod}{A model object (or data.frame). For the moment only lm and a plain dataset is implemented}

\item{...}{not used}
}
\value{
a matrix with for each variable de variance inflation factor, the degrees of freedom  en the rescaled variance inflation factor based on the degrees of freedom.
}
\description{
Variance inflation factor

Variance inflaction factor (lm model)

Variance inflaction factor (data.frame)

Variance inflation factor (default routine)
}
\examples{
{
vif(airquality)
}
}
