% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sessioninfo.R
\name{write_sessioninfo}
\alias{write_sessioninfo}
\title{Write essential session info to a file or connection}
\usage{
write_sessioninfo(file, pkgs = NULL, include_base = FALSE)
}
\arguments{
\item{file}{A file path or a \code{\link[base]{connection}}.}

\item{pkgs}{Either a vector of package names or NULL. If \code{NULL},
displays all loaded packages. If a character vector, also, includes
all dependencies of the package.}

\item{include_base}{Include base packages in summary? By default this is
false since base packages should always match the R version.}
}
\description{
Writes essentials of
\code{\link[sessioninfo:session_info]{sessioninfo::session_info()}}
to a file or connection in order to
maintain versionable documentation of a workflow and aid
reproducibility.
}
\details{
The following information is written:
\itemize{
\item{essential platform info: version, os, system, ctype}
\item{essential info on loaded packages: package, loadedversion, date,
source.}
}
}
\examples{
\dontrun{
write_sessioninfo(file = "./sessioninfo.yml")
}
}
