% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_sqlite.R
\name{append_to_sqlite}
\alias{append_to_sqlite}
\title{Callback function that appends new sections to the SQLite table.}
\usage{
append_to_sqlite(con, table_name, date_cols, datetime_cols)
}
\arguments{
\item{con}{A valid connection to SQLite database.}

\item{table_name}{Name of the table to store the data table in the sqlite
database.}

\item{date_cols}{Name of columns containing Date objects}

\item{datetime_cols}{Name of columns containint POSIXt objects.}

\item{x}{Data.frame we are reading from.}
}
\description{
Callback function that appends new sections to the SQLite table.
}
\keyword{internal}
