% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_sqlite.R
\name{csv_to_sqlite}
\alias{csv_to_sqlite}
\title{Save a delimited text table into a single table sqlite database}
\usage{
csv_to_sqlite(csv_file, sqlite_file, table_name, delim = ",",
  pre_process_size = 1000, chunk_size = 50000,
  show_progress_bar = TRUE, ...)
}
\arguments{
\item{csv_file}{name of the text file to convert}

\item{sqlite_file}{name of the newly created sqlite file}

\item{table_name}{name of the table to store the data table in the sqlite
dbase}

\item{delim}{text file delimiter (default ",")}

\item{pre_process_size}{the number of lines to check the data types of the
individual columns (default 1000)}

\item{chunk_size}{the number of lines to read for each chunk (default 50000)}

\item{show_progress_bar}{show progress bar (default TRUE)}

\item{...}{Further arguments to be passed to \code{read_delim}.}

\item{callback}{A callback function to call on each chunk.}
}
\value{
a SQLite database
}
\description{
The table can be a comma separated (csv) or a tab separated (tsv) or any
other delimited text file. The file is read in chunks. Each chunk is copied
in the same sqlite table database before the next chunk is loaded into
memory.
}
