% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_sqlite.R
\name{csv_to_sqlite}
\alias{csv_to_sqlite}
\title{Save a single CSV-table into a single table sqlite database}
\usage{
csv_to_sqlite(csv_file, sqlite_file, table_name, delim = ",",
  pre_process_size = 1000, chunk_size = 50000)
}
\arguments{
\item{csv_file}{name of the CSV file to convert}

\item{sqlite_file}{name of the newly created sqlite file}

\item{table_name}{name of the table to store the data table in the sqlite
dbase}

\item{delim}{text file delimiter (default ",")}

\item{pre_process_size}{the number of lines to check the data types of the
individual columns (default 1000)}

\item{chunk_size}{the number of lines to read for each chunk (default 50000)}
}
\value{
a SQLite database
}
\description{
Save a single CSV-table into a single table sqlite database
}
