% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_knmi_data_hour.R
\name{download_knmi_data_hour}
\alias{download_knmi_data_hour}
\title{Download KNMI hourly data to file}
\usage{
download_knmi_data_hour(
  stations,
  variables,
  start_date,
  end_date,
  output_file = "knmi_download.txt"
)
}
\arguments{
\item{stations}{list of integers. For an overview,
see https://www.daggegevens.knmi.nl/klimatologie/uurgegevens}

\item{variables}{list of variables, options are:
\itemize{
\item{"WIND = DD:FH:FF:FX"}{Wind}
\item{"TEMP = T:T10N:TD"}{Temperatuur}
\item{"SUNR = SQ:Q"}{Zonneschijnduur en globale straling}
\item{"PRCP = DR:RH"}{Neerslag en potentiële verdamping}
\item{"VICL = VV:N:U"}{Zicht, bewolking en relatieve vochtigheid}
\item{"WEER = M:R:S:O:Y:WW"}{Weerverschijnselen, weertypen}
\item{"ALL"}{alle variabelen}
}}

\item{start_date}{date interpretable string to define start of the period}

\item{end_date}{date interpretable string to define end of the period}

\item{output_file}{path/filename of the output_file_name}
}
\value{
response containing, status_code, content,...
}
\description{
Meer info, zie
https://www.knmi.nl/kennis-en-datacentrum/achtergrond/data-ophalen-vanuit-een-script
}
\examples{
\dontrun{
download_knmi_data_hour(c(310, 319), "PRCP", "2011-01-01", "2012-02-17",
  output_file = "knmi_output_download.txt"
)
}
}
\seealso{
Other download_functions: 
\code{\link{download_zenodo}()},
\code{\link{extract_soil_map_data}()},
\code{\link{read_kmi_data}()},
\code{\link{read_kml_file}()},
\code{\link{read_knmi_data}()},
\code{\link{read_mow_data}()}
}
\concept{download_functions}
