% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_zenodo.R
\name{download_zenodo}
\alias{download_zenodo}
\title{Get data from a Zenodo archive}
\usage{
download_zenodo(doi, path = ".")
}
\arguments{
\item{doi}{a doi pointer to the Zenodo archive starting with '10.5281/zenodo.'. See examples.}

\item{path}{Path where the data must be downloaded.
Defaults to the working directory.}
}
\description{
This function will download an entire archive from Zenodo (\url{https://zenodo.org}).
It only works for Zenodo created DOI (not when the DOI is for
example derived from Zookeys.)
}
\examples{
\dontrun{
# Example download of an archive containing a single zip
download_zenodo(doi = "10.5281/zenodo.1283345")
# Example download of an archive containing multiple files
# (all files will be downloaded)
download_zenodo(doi = "10.5281/zenodo.1172801")
# Example download of an archive containing a single pdf file
download_zenodo(doi = "10.5281/zenodo.168478")
}
}
\seealso{
Other download: 
\code{\link{download_knmi_data_hour}()}
}
\concept{download}
