% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_zenodo.R
\name{download_zenodo}
\alias{download_zenodo}
\title{Get data from a Zenodo archive}
\usage{
download_zenodo(doi, path = ".", parallel = FALSE, quiet = FALSE)
}
\arguments{
\item{doi}{a doi pointer to the Zenodo archive starting with '10.5281/zenodo.'. See examples.}

\item{path}{Path where the data must be downloaded.
Defaults to the working directory.}

\item{parallel}{Logical (\code{FALSE} by default).
If \code{TRUE}, will run a number of parallel processes, each downloading
another file.
This is useful when multiple large files are present in the Zenodo
record, which otherwise would be downloaded sequentially.
Of course, the operation is limited by bandwidth and traffic limitations.}

\item{quiet}{Logical (\code{FALSE} by default).
Do you want to suppress informative messages (not warnings)?}
}
\description{
This function will download an entire archive from Zenodo (\url{https://zenodo.org}).
It only works for Zenodo created DOI (not when the DOI is for
example derived from Zookeys.)
}
\examples{
\dontrun{
# Example download of an archive containing a single zip
download_zenodo(doi = "10.5281/zenodo.1283345")
download_zenodo(doi = "10.5281/zenodo.1283345", quiet = TRUE)
# Example download of an archive containing multiple files
# using parallel download
# (multiple files will be simultaneously downloaded)
download_zenodo(doi = "10.5281/zenodo.1172801", parallel = TRUE)
# Example download of an archive containing a single pdf file
download_zenodo(doi = "10.5281/zenodo.168478")
}
}
\seealso{
Other download functions: 
\code{\link{download_knmi_data_hour}()}
}
\concept{download functions}
