% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soil_map_data.R
\name{extract_soil_map_data}
\alias{extract_soil_map_data}
\title{Extract soil properties from Flemish soil map}
\usage{
extract_soil_map_data(x_lam, y_lam, properties_of_interest)
}
\arguments{
\item{x_lam}{The numeric value of the X coordinate in CRS
'Belge 1972 / Belgian Lambert 72' (EPSG:31370).}

\item{y_lam}{The numeric value of the Y coordinate in CRS
'Belge 1972 / Belgian Lambert 72' (EPSG:31370).}

\item{properties_of_interest}{A vector or properties, as a subset of these
provided by the webservice (see the examples section for how to obtain the
list). Default Bodemserie, Unibodemtype and Bodemtype.}
}
\value{
A data.frame with the properties as column headers.
}
\description{
This function queries the
\href{https://www.geopunt.be/catalogus/datasetfolder/5c129f2d-4498-4bc3-8860-01cb2d513f8f}{the Flemish soil map}
attributes at a given coordinate,
by using the affiliated WFS service provided by DOV. The user can pick the
properties of interest. See the examples section to see how to obtain a full
list of possible properties of interest.
Coordinates should be given in the 'Belge 1972 / Belgian Lambert 72'
(\href{https://epsg.io/31370}{EPSG:31370}) coordinate reference system.
When outside the Flemish region, an NA value is given for each of the properties.
}
\examples{
\dontrun{
library(inborutils)
extract_soil_map_data(173995.67, 212093.44)

# code to obtain list of all possible soil properties
library(ows4R)
library(purrr)
wfs <- "https://www.dov.vlaanderen.be/geoserver/bodemkaart/bodemtypes/wfs"
wfs_client <- WFSClient$new(wfs,
  serviceVersion = "1.1.0"
)
wfs_client$
  getCapabilities()$
  findFeatureTypeByName("bodemkaart:bodemtypes")$
  getDescription() \%>\%
  map_chr(function(x) {
    x$getName()
  })

extract_soil_map_data(173995.67, 212093.44,
  properties_of_interest = c(
    "Eenduidige_legende",
    "Textuurklasse"
  )
)
}
}
\seealso{
Other download_functions: 
\code{\link{download_knmi_data_hour}()},
\code{\link{download_zenodo}()},
\code{\link{read_kmi_data}()},
\code{\link{read_kml_file}()},
\code{\link{read_knmi_data}()},
\code{\link{read_mow_data}()}
}
\concept{download_functions}
