% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/florabank_queries.R
\name{florabank_observations}
\alias{florabank_observations}
\title{Get all validated observations for a taxon from the florabank database}
\usage{
florabank_observations(connection, scient_name, dutch_name,
  collect = FALSE)
}
\arguments{
\item{connection}{A connection to the florabank database. See the example section
for how to connect and disconnect to the database.}

\item{scient_name}{(Part of) a scientific taxon name}

\item{dutch_name}{(Part of) a Dutch taxon name, may contain wildcards}

\item{collect}{If FALSE (the default), a remote tbl object is returned. This
is like a reference to the result of the query but the full result of the
query is not brought into memory. If TRUE the full result of the query is
collected (fetched) from the database and brought into memory of the working
environment.}
}
\value{
A dataframe with the following variables: "NaamNederlands",
"NaamWetenschappelijk", "Bron", "BeginDatum", "EindDatum", "hok",
"Toponiem", "CommentaarTaxon", "CommentaarHabitat", "MedewerkerID",
"WaarnemingID", "X_waarneming", "Y_waarneming", "X_meting", "Y_meting"
}
\description{
This function takes as input either (part of) a scientific name or (part of)
the Dutch name for a taxon, queries the florabank, and returns a dataframe
observation level information about the matching taxa. The following
wildcards may be used within the input parameters:
\code{\%}: Represents zero or more characters;
\code{_}: Represents a single character;
\code{[]}: Represents any single character within the brackets, e.g. \code{[sp]}
\code{^}: Represents any character not in the brackets, e.g. \code{[^sp]}
\code{-}: Represents a range of characters, e.g. \code{[a-z]}
}
\examples{
\dontrun{
# connect to florabank
db_connectie <- connect_inbo_dbase("D0021_00_userFlora")

# query and collect the data using scientific name
Dactmacu1 <-	florabank_observations(db_connectie,
scient_name = 'Dactylorhiza maculata (L.) Soó', collect = TRUE)

# the same species but using Dutch name
Dactmacu2 <-	florabank_observations(db_connectie,
dutch_name = 'Gevlekte orchis', collect = TRUE)

all.equal(Dactmacu1, Dactmacu2)

# use wildcards to retrieve all partial matches to a name
# using default for collect will return a lazy query
Dactmacu3 <-	florabank_observations(db_connectie,
scient_name = 'Dactylorhiza maculata\%')

# to collect the data for a lazy query you can also do:
Dactmacu3 <- collect(Dactmacu3)

# disconnect from florabank
dbDisconnect(db_connectie)
}
}
\seealso{
Other florabank: \code{\link{florabank_taxon_ifbl_year}},
  \code{\link{florabank_traits}}
}
\concept{florabank}
