% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_species_name_match.R
\name{gbif_species_name_match}
\alias{gbif_species_name_match}
\title{Add species information provided by the GBIF taxonomic backbone API}
\usage{
gbif_species_name_match(df, name_col, gbif_terms = c("usageKey",
  "scientificName", "rank", "order", "matchType", "phylum", "kingdom",
  "genus", "class", "confidence", "synonym", "status", "family"))
}
\arguments{
\item{df}{data.frame with species information}

\item{name_col}{char column name of the column containing the scientific
names used for the name matching with the GBIF taxonomic backbone.}

\item{gbif_terms}{list of valid GBIF terms to add as additional columns to
the data.frame}
}
\value{
df with GBIF information as additional columns
}
\description{
This functions extends an existing dataframe with additional columns provided
by the GBIF taxonomic backbone and matched on the species scientific name,
which need to be an available column in the dataframe.
}
\details{
This function is essentially a wrapper around the existing rgbif
functionality `name_lookup()` and extends the application to a data.frame.
For more information on the name matching API of GBIF on which rgbif relies,
see https://www.gbif.org/developer/species
}
