% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_projection.R
\name{guess_projection}
\alias{guess_projection}
\title{Take a dataset and plot in different projections on a leaflet map to
check the expected projection}
\usage{
guess_projection(
  df,
  col_long,
  col_lat,
  belgium = TRUE,
  projections = c("epsg:4326", "epsg:31370", "epsg:28992", "epsg:32631", "epsg:3812",
    "epsg:3035")
)
}
\arguments{
\item{df}{data.frame with a x and y coordinate column}

\item{col_long}{(char) name of the x (longitude) column}

\item{col_lat}{(char) name of the y (latitude) column}

\item{belgium}{(bool) If TRUE, coordinates are expected to be in Belgium}

\item{projections}{(list) epsg codes of the different projections to evaluate
By default, the following six projections are added to the map:
\itemize{
\item{"epsg:4326"}{WGS 84,
http://spatialreference.org/ref/epsg/wgs-84/}
\item{"epsg:31370"}{Belge 1972/Belgian Lambert 72,
http://spatialreference.org/ref/epsg/31370/}
\item{"epsg:28992"}{Amersfoort/Rijksdriehoek nieuw,
http://spatialreference.org/ref/epsg/28992/}
\item{"epsg:32631"}{WGS 84 / UTM zone 31N,
http://spatialreference.org/ref/epsg/32631/}
\item{"epsg:3812"}{ETRS89/Belgian Lambert 2008,
http://spatialreference.org/ref/epsg/3812/}
\item{"epsg:3035"}{ETRS89 / ETRS-LAEA,
http://spatialreference.org/ref/epsg/3035/}
}}
}
\description{
When retrieving a number of data-points, but the kind of projection is not
provided/known (for any reason), this utility plots the data for different
projections on a map to make comparison possible. The main function is
\code{guess_projection } which creates the map with different options.
Custom projections can be provided by the user as well.
}
\details{
For each of the given projections, the coordinates columns are given the
the specified projection. In a next step, these projections are converted to
wgs84 and plotted on a openstreetmap background with leaflet. The interactive
map provides the possibility to select/unselect specific layers.
}
\examples{
\dontrun{
guess_projection(data, "x", "y")
}

}
