% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inborutils_defunct.R
\name{inborutils-defunct}
\alias{inborutils-defunct}
\title{Defunct functions in package \pkg{inborutils}}
\description{
The functions listed below are moved to package \pkg{inbodb}
and removed from \pkg{inborutils}. Please install package
\pkg{inbodb} to continue using these functions (instructions on
https://inbo.github.io/inbodb/).
}
\keyword{internal}
