% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inborutils.R
\docType{package}
\name{inborutils-package}
\alias{inborutils}
\alias{inborutils-package}
\title{inborutils: Collection of Useful R Utilities}
\description{
While working on research projects, typical small functionalities 
    are useful across these projects. Instead of copy-pasting these functions in 
    all indidivual project repositories/folders, this package collects these 
    functions for reuse by ourself and - if useful - others as well.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/inbo/inborutils}
  \item \url{https://inbo.github.io/inborutils}
  \item Report bugs at \url{https://github.com/inbo/inborutils/issues}
}

}
\author{
\strong{Maintainer}: Hans Van Calster \email{hans.vancalster@inbo.be} (\href{https://orcid.org/0000-0001-8595-8426}{ORCID})

Authors:
\itemize{
  \item Damiano Oldoni \email{damiano.oldoni@inbo.be} (\href{https://orcid.org/0000-0003-3445-7562}{ORCID})
  \item Stijn Van Hoey \email{stijnvanhoey@gmail.com} (\href{https://orcid.org/0000-0001-6413-3185}{ORCID})
}

Other contributors:
\itemize{
  \item Dimitri Brosens \email{dimitri.brosens@inbo.be} (\href{https://orcid.org/0000-0002-0846-9116}{ORCID}) [contributor]
  \item Peter Desmet \email{peter.desmet@inbo.be} (\href{https://orcid.org/0000-0002-8442-8025}{ORCID}) [contributor]
  \item Frank Huysentruyt \email{frank.huysentruyt@inbo.be} (\href{https://orcid.org/0000-0002-3071-9126}{ORCID}) [contributor]
  \item Thierry Onkelinx \email{thierry.onkelinx@inbo.be} (\href{https://orcid.org/0000-0001-8804-4216}{ORCID}) [contributor]
  \item Floris Vanderhaeghe \email{floris.vanderhaeghe@inbo.be} (\href{https://orcid.org/0000-0002-6378-6229}{ORCID}) [contributor]
  \item Pieter Verschelde \email{pieter.verschelde@inbo.be} (\href{https://orcid.org/0000-0002-9199-421X}{ORCID}) [contributor]
  \item Els De Bie \email{els.debie@inbo.be} (\href{https://orcid.org/0000-0001-7679-743X}{ORCID}) [contributor]
  \item Research Institute for Nature and Forest (INBO) \email{info@inbo.be} [copyright holder]
}

}
\keyword{internal}
