% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inboveg_classification.R
\name{inboveg_classification}
\alias{inboveg_classification}
\title{Provide vegetation classifications}
\usage{
inboveg_classification(connection, survey_name)
}
\arguments{
\item{connection}{odb DBIConnection-class, using DBI-ODBC like connectors}

\item{survey_name}{vector of char or NULL (default NULL),
i.e. all surveys are taken into account}
}
\value{
data.frame
}
\description{
Get the available vegatation classifications for the vegetation surveys
}
\examples{
\dontrun{
con <- connect_inbo_dbase("D0010_00_Cydonia")
classifications <- inboveg_classification(con,
                           survey_name = c('Sigma_LSVI_2012'))
dbDisconnect(con)
}
}
