% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inboveg_survey.R
\name{inboveg_survey}
\alias{inboveg_survey}
\title{Survey overview}
\usage{
inboveg_survey(connection)
}
\arguments{
\item{connection}{odb DBIConnection-class, using DBI-ODBC like connectors}
}
\value{
data.frame
}
\description{
Survey overview
}
\examples{
\dontrun{
library(DBI)
con <- dbConnect(odbc::odbc(), dsn="Cydonia-prd")
surveys <- inboveg_coverage(con)
}
}
