% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inboveg_synonym.R
\name{inboveg_synonym_list}
\alias{inboveg_synonym_list}
\title{Provide accepted names list from INBOVEG database}
\usage{
inboveg_synonym_list(connection, TaxonListGIVID = "TL2011092815101010",
  ListName = "INBO-2011 Sci")
}
\arguments{
\item{connection}{odb}

\item{TaxonListGIVID}{char}

\item{ListName}{char}
}
\value{
DBIConnection-class, using DBI-ODBC like connectors
}
\description{
Get all taxon names available from an INBOVEG database taxon list
(identified with the `TaxonListGIVID`) and give for each name the corresponding
preferred taxon name as available in the specified reference list
(identified by `ListName`)
}
\examples{
\dontrun{
library(DBI)
con <- dbConnect(odbc::odbc(), dsn="Cydonia-prd")
plant_overview <- inboveg_synonym_list(con,
                           TaxonListGIVID = 'TL2011092815101010',
                           ListName = 'INBO-2011 Sci')
}
}
