% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_inbo_dbase.R
\name{on_connection_opened}
\alias{on_connection_opened}
\title{Rstudio Viewer integration}
\usage{
on_connection_opened(connection, code, type)
}
\arguments{
\item{connection}{odbc connection}

\item{code}{dbase connection code}

\item{server}{INBO database server name}
}
\description{
See https://stackoverflow.com/questions/48936851/calling-odbc-connection-within-function-does-not-display-in-rstudio-connection and https://rstudio.github.io/rstudio-extensions/connections-contract.html#persistence
}
