% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rain_knmi_2012}
\alias{rain_knmi_2012}
\title{Example data.frame with KNMI downloaded data}
\format{
A data frame with 1536 rows and 9 variables:
\itemize{
\item{value: measured value}
\item{datetime: datetime of the measurement}
\item{unit: unit (mm)}
\item{variable_name: precipitation}
\item{longitude: coordinate}
\item{latitude: coordinate}
\item{location_name: station name}
\item{source_filename: filename from which the data was read}
\item{quality_code: empty string as KNMI does not provide this}
}
}
\usage{
rain_knmi_2012
}
\description{
A dataset containing the rainfall from January 1st till February 1st for
Vlissingen and Westdorpe, as downloaded from KNMI
}
\keyword{datasets}
