% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kmi_data.R
\name{read_kmi_data}
\alias{read_kmi_data}
\title{Load \code{KMI} data as data.frame}
\usage{
read_kmi_data(filename, n_max = Inf)
}
\arguments{
\item{filename}{path/filename of the \code{KMI} datafile}

\item{n_max}{integer; shortcut for testing, loading just a section of the
data}
}
\value{
\code{data.frame} with the headers \code{location_name}, \code{datetime} (UTC),
\code{value}, \code{unit}, \code{variable_name}, \code{latitude}, \code{longitude} and
\code{source_filename}
}
\description{
Load \code{KMI} data as data.frame
}
\examples{
\dontrun{
# see vignettes for more examples
file_path <- "kmi_file.txt"
precipation_kmi <- read_kmi_data(file_path)
}
}
\seealso{
Other download_functions: 
\code{\link{download_knmi_data_hour}()},
\code{\link{download_zenodo}()},
\code{\link{extract_soil_map_data}()},
\code{\link{read_kml_file}()},
\code{\link{read_knmi_data}()},
\code{\link{read_mow_data}()}
}
\concept{download_functions}
