% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kml_file.R
\name{read_kml_file}
\alias{read_kml_file}
\title{Extract date and coordinate information from Google Maps kml file}
\usage{
read_kml_file(filename)
}
\arguments{
\item{filename}{(char) filename/path of the kml file}
}
\value{
df with columns datetime, x (numeric), y (numeric)
}
\description{
Extract date and coordinate information from Google Maps kml file
}
\examples{
\dontrun{
df <- read_kml_file("BE1002.kml")
}
}
\seealso{
Other download functions: 
\code{\link{download_knmi_data_hour}()},
\code{\link{download_zenodo}()},
\code{\link{extract_soil_map_data}()},
\code{\link{read_kmi_data}()},
\code{\link{read_knmi_data}()},
\code{\link{read_mow_data}()}
}
\concept{download functions}
