% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_coordinates.R
\name{reproject_coordinates}
\alias{reproject_coordinates}
\title{Transform x-y coordinates from data.frame columns.}
\usage{
reproject_coordinates(df, col_long, col_lat, crs_input, crs_output)
}
\arguments{
\item{df}{A data.frame with a x and y coordinate column.}

\item{col_long, col_lat}{Column names or positions of the x and y
column. They are passed to \code{\link[tidyselect]{vars_pull}}.
These arguments are passed by expression and support
\code{\link{quasiquotation}}
(you can unquote column names or column positions).}

\item{crs_input}{Projection string of class CRS-class (\code{sp} compatible)
or crs-class (\code{sf} compatible) defining the current CRS.}

\item{crs_output}{Projection string of class CRS-class (\code{sp} compatible)
or crs-class (\code{sf} compatible) defining the CRS to convert to.}
}
\value{
A data.frame with the same columns, but transformed coordinates for
the x and y column values.
}
\description{
This function extracts x-y coordinates from a data.frame by means of the
given coordinate reference system (CRS), transforms them to the new CRS and
assign them back to the data.frame columns.
}
\seealso{
Other GIS_utilities: 
\code{\link{guess_projection}()},
\code{\link{plot_coordinates_on_map}()},
\code{\link{transform_coordinates}()}
}
\concept{GIS_utilities}
