% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_example}
\alias{species_example}
\title{Example data.frame with species name column}
\format{A data frame with 53940 rows and 10 variables:
\describe{
\item{scientificName}{scientific name of the species}
\item{kingdom}{kingdom to which the species belongs}
\item{euConcernStatus}{level of concern according to EU directives}
}}
\usage{
species_example
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
