% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_example}
\alias{species_example}
\title{Example data.frame with species name column}
\format{A data frame with 3 rows and 3 variables
\itemize{
\item {speciesName: name of the species}
\item {kingdom: kingdom to which the species belongs}
\item {euConcernStatus: level of concern according to EU directives}
}}
\usage{
species_example
}
\description{
A dataset containing 3 taxa to be matched with GBIF Taxonomy Backbone. The
variable are as follows:
}
\keyword{datasets}
