% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sessioninfo.R
\name{write_sessioninfo}
\alias{write_sessioninfo}
\title{Write essential session info to a file or connection}
\usage{
write_sessioninfo(file)
}
\arguments{
\item{file}{A file path or a \code{\link[base]{connection}}.}
}
\description{
Writes essentials of
\code{\link[sessioninfo:session_info]{sessioninfo::session_info()}}
to a file or connection in order to
maintain versionable documentation of a workflow and aid
reproducibility.
}
\details{
The following information is written:
\itemize{
\item{essential platform info: version, os, system, ctype}
\item{essential info on loaded packages: package, loadedversion, date,
source if different from CRAN.}
}
}
\examples{
\dontrun{
write_sessioninfo(file = "./sessioninfo.txt")
}
}
